+++ packages/svelte-table/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @layerstack/svelte-table
 
+## 1.0.1-next.3
+
+### Patch Changes
+
+- Updated dependencies [[`4ed65a7`](https://github.com/techniq/layerstack/commit/4ed65a76562db9af2d18a196a2ba9e58f959aa5c)]:
+  - @layerstack/utils@2.0.0-next.3
+  - @layerstack/svelte-actions@1.0.1-next.3
+
+## 1.0.1-next.2
+
+### Patch Changes
+
+- Updated dependencies [[`955e07b`](https://github.com/techniq/layerstack/commit/955e07b5aed62acd8afba10f9eaa68b90d72bb74)]:
+  - @layerstack/utils@1.1.0-next.2
+  - @layerstack/svelte-actions@1.0.1-next.2
+
+## 1.0.1-next.1
+
+### Patch Changes
+
+- Update dependencies ([#31](https://github.com/techniq/layerstack/pull/31))
+
+- Updated dependencies [[`b7ca80b`](https://github.com/techniq/layerstack/commit/b7ca80b6a8a07c53ec4a99864ec6b9fd1ecab0b4)]:
+  - @layerstack/svelte-actions@1.0.1-next.1
+  - @layerstack/utils@1.1.0-next.1
+
+## 1.0.1-next.0
+
+### Patch Changes
+
+- Updated dependencies [[`6cdcb26`](https://github.com/techniq/layerstack/commit/6cdcb26cabd0539cee2885efbaa39e7787a34114), [`c6bb443`](https://github.com/techniq/layerstack/commit/c6bb443d12ee12bd69417dcfa5880b6bf78c9f09), [`c13e65e`](https://github.com/techniq/layerstack/commit/c13e65e059d690cbc282635f6e48a27d715e5997)]:
+  - @layerstack/utils@1.1.0-next.0
+  - @layerstack/svelte-actions@1.0.1-next.0
+
 ## 1.0.1
 
 ### Patch Changes