+++ packages/utils/CHANGELOG.md
@@ -1,10 +1,40 @@
 # @layerstack/utils
 
+## 2.0.0-next.3
+
+### Major Changes
+
+- breaking: Replace `getDuration()` / `humanizeDuration()` utils with `Duration` class (with `.format()` method) ([#62](https://github.com/techniq/layerstack/pull/62))
+
+## 1.1.0-next.2
+
+### Patch Changes
+
+- fix(format): Do not truncate decimal values when using `metric` preset (ex. 0.5) ([#53](https://github.com/techniq/layerstack/pull/53))
+
 ## 1.0.1
 
 ### Patch Changes
 
 - fix(format): Do not truncate decimal values when using `metric` preset (ex. 0.5) ([#53](https://github.com/techniq/layerstack/pull/53))
+
+## 1.1.0-next.1
+
+### Patch Changes
+
+- Update dependencies ([#31](https://github.com/techniq/layerstack/pull/31))
+
+## 1.1.0-next.0
+
+### Minor Changes
+
+- Tailwind 4 support ([#32](https://github.com/techniq/layerstack/pull/32))
+
+### Patch Changes
+
+- feat: Add `mapKeys()` and `mapValues()` object utils ([#32](https://github.com/techniq/layerstack/pull/32))
+
+- feat: Add `NestedRecord` type ([#32](https://github.com/techniq/layerstack/pull/32))
 
 ## 1.0.0